requirements.txt
# LHC Data Pipeline - Python Dependencies

kafka-python>=2.0.2
flask>=3.0.0
flask-socketio>=5.3.0
